+++ comp2013/src/main/java/comp2013/SnakeMusic.java
@@ -10,7 +10,7 @@ public class SnakeMusic implements IMusic
 	private String M_FileName;
 	private Media M_MusicMedia;
 	private MediaPlayer M_MediaPlayer;
-	private float M_CurrentVolume;
+	private double M_CurrentVolume;
 
 	public SnakeMusic(String filename)
 	{
@@ -18,14 +18,11 @@ public class SnakeMusic implements IMusic
 		this.M_MusicMedia = new Media(M_FileName);
 		this.M_MediaPlayer = new MediaPlayer(M_MusicMedia);
 		this.M_CurrentVolume = 1;
-		System.out.printf("Here1\n");
-
 	}
 
 	// Sets the music that you want to play
 	@Override
 	public void setMusic(String filename) {
-		System.out.printf("Here2\n");
 
 		if(filename != null) {
 			// Set the filename.
@@ -48,12 +45,11 @@ public class SnakeMusic implements IMusic
 	}
 
 	@Override
-	public void setVolume(float volume){
+	public void setVolume(double volume){
 		// Update the volume variable
 		this.M_CurrentVolume = volume;
 		// Set the new volume of the player
 		M_MediaPlayer.setVolume(M_CurrentVolume);
-		System.out.printf("Here3\n");
 
 	}
 	@Override
@@ -87,8 +83,6 @@ public class SnakeMusic implements IMusic
 		if (this.M_MediaPlayer != null) {
 			// Enable looping if true.
 			if(loop) {
-				System.out.printf("Here4\n");
-
 				this.M_MediaPlayer.setOnEndOfMedia(() -> {
 					// When the end of media is reached,
 					// reset the playback position to the beginning
@@ -97,8 +91,6 @@ public class SnakeMusic implements IMusic
 			}
 			// Disable looping if false
 			else{
-				System.out.printf("Here5\n");
-
 				this.M_MediaPlayer.setOnEndOfMedia(null); // Disable looping
 			}
 		}
@@ -109,8 +101,6 @@ public class SnakeMusic implements IMusic
 	public void playMusic() {
 		if(M_MediaPlayer.getStatus() != MediaPlayer.Status.PLAYING){
 			M_MediaPlayer.play();
-			System.out.printf("Here6\n");
-
 		}
 	}
 
@@ -122,8 +112,6 @@ public class SnakeMusic implements IMusic
 		Duration seekTime = Duration.seconds(timeStamp);
 		M_MediaPlayer.seek(seekTime);
 		M_MediaPlayer.play();
-		System.out.printf("Here7\n");
-
 	}
 
 	// Stops the music from playing
@@ -131,8 +119,6 @@ public class SnakeMusic implements IMusic
 	public void pauseMusic() {
 		if(M_MediaPlayer.getStatus() == MediaPlayer.Status.PLAYING){
 			M_MediaPlayer.pause();
-			System.out.printf("Here8\n");
-
 		}
 	}
 
@@ -142,8 +128,6 @@ public class SnakeMusic implements IMusic
 			M_MediaPlayer.pause();
 			// Remove the previous object
 			M_MediaPlayer.dispose();
-			System.out.printf("Here9\n");
-
 		}
 	}