+++ docs/changelog.md
@@ -146,3 +146,10 @@
 - Added methods to SnakeMusic, "setVolume, volumeMute, increaseVolume, decreaseVolume, pauseMusic".
 - Reworked stopMusic so that it disposes of the old media player once done.
 - Fixed a bug with restartGame where it would interfere with the menu music.
+
+### 11/12/23
+- Reworked how the controller restartGame method works. No longer creates a new model.
+- Removed the finished variable from model.
+- Removed getters and setters for finished variable from model.
+- Added a slider that changes the volume to the settings page.
+- Updated setVoulme type to double from float.